Fix ios builds by not trying to compile go-ps in ios builds
