+++ go/install/fuse_status_default.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Keybase, Inc. All rights reserved. Use of
 // this source code is governed by the included BSD license.
 
-// +build !darwin,!windows
+// +build !darwin,!windows ios
 
 package install
