+++ go/install/terminate_mobile.go
@@ -1,0 +1,8 @@
+// +build ios android
+
+package install
+
+// TerminateApp will stop the Keybase (UI) app
+func TerminateApp(context Context, log Log) error {
+	return nil
+}
